refactor(fmt): Use deno fmt to clean code
To complete the test workflow.
