+++ src/libs/mod.ts
@@ -1,2 +1,1 @@
 export * from "./io.ts";
-export * from "./convert.ts";
